python3/flask-projects/flask-requirements.txt
Package                Version   
---------------------- ----------
alembic                1.0.9     
appdirs                1.4.3     
asn1crypto             0.24.0    
attrs                  19.1.0    
autopep8               1.4.4     
backcall               0.1.0     
black                  19.3b0    
certifi                2018.11.29
cffi                   1.12.3    
chardet                3.0.4     
Click                  7.0       
cryptography           2.6.1     
decorator              4.4.0     
entrypoints            0.3       
flake8                 3.7.7     
Flask                  1.0.2     
Flask-Login            0.4.1     
Flask-Migrate          2.4.0     
Flask-SQLAlchemy       2.3.2     
Flask-WTF              0.14.2    
future                 0.17.1    
greenlet               0.4.15    
idna                   2.8       
ipython                7.4.0     
ipython-genutils       0.2.0     
isort                  4.3.17    
itsdangerous           1.1.0     
jedi                   0.13.3    
Jinja2                 2.10.1    
Mako                   1.0.9     
MarkupSafe             1.1.1     
mccabe                 0.6.1     
msgpack                0.6.1     
neovim                 0.3.1     
parso                  0.4.0     
pexpect                4.7.0     
pickleshare            0.7.5     
Pillow                 6.0.0     
pip                    19.0.3    
pluggy                 0.9.0     
prompt-toolkit         2.0.9     
ptyprocess             0.6.0     
pycodestyle            2.5.0     
pycosat                0.6.3     
pycparser              2.19      
pydocstyle             3.0.0     
pyflakes               2.1.1     
Pygments               2.3.1     
pynvim                 0.3.2     
pyOpenSSL              19.0.0    
PySocks                1.6.8     
python-dateutil        2.8.0     
python-editor          1.0.4     
python-jsonrpc-server  0.1.2     
python-language-server 0.26.1    
requests               2.21.0    
rope                   0.14.0    
ruamel-yaml            0.15.46   
setuptools             41.0.0    
six                    1.12.0    
snowballstemmer        1.2.1     
SQLAlchemy             1.3.3     
toml                   0.10.0    
traitlets              4.3.2     
urllib3                1.24.2    
wcwidth                0.1.7     
Werkzeug               0.15.2    
wheel                  0.33.1    
WTForms                2.2.1     
yapf                   0.27.0    
